requirements.txt
pymobiledevice3==4.20.11
PyYAML==6.0.1
geopy==2.4.1
customtkinter==5.2.2